refactor(reader): extract annotation bridge payloads
Why:
- BibleReaderController remained over 9k lines after the speech extraction, so #146 still needed a materially larger responsibility split.
- Bookmark, label, My Notes, and StudyPad bridge payload projection was embedded in the controller despite being a distinct reader bridge contract.

What Changed:
- Added BibleReaderAnnotationPayloadFactory to own bookmark, label, StudyPad, and My Notes DTO projection.
- Moved reader bridge document and update payload DTOs into BibleReaderBridgePayloads.
- Replaced controller payload-builder implementations with thin delegating wrappers while preserving existing bridge method names and payload shapes.

Validation:
- python3 scripts/check_repo_standards.py docblocks
- DEVELOPER_DIR=/Applications/Xcode.app/Contents/Developer xcodebuild -project AndBible.xcodeproj -scheme AndBible -configuration Debug -destination id=0BFFE049-A46B-4BCA-A6B3-281EA931DD7F -derivedDataPath .derivedData build
- DEVELOPER_DIR=/Applications/Xcode.app/Contents/Developer xcodebuild -project AndBible.xcodeproj -scheme AndBible -configuration Debug -destination id=0BFFE049-A46B-4BCA-A6B3-281EA931DD7F -derivedDataPath .derivedData -only-testing:AndBibleTests/AndBibleTests/testReaderBookmarkBridgeUpdateEmitsTypedPayloadShape -only-testing:AndBibleTests/AndBibleTests/testReaderStudyPadDocumentBridgeEmissionUsesTypedNestedPayloads -only-testing:AndBibleTests/AndBibleTests/testReaderConfigPayloadNormalizesInvalidPageScrollAmount -only-testing:AndBibleTests/AndBibleTests/testBridgeReadingProgressMessagesMutateNativeStoreAndEmitCounts -only-testing:AndBibleTests/AndBibleTests/testBridgeReadingProgressSettingsAndPresentationHandoffs -only-testing:AndBibleTests/AndBibleTests/testDefinitionDocumentRequestedBeforeClientReadyReplaysAfterClientReady -only-testing:AndBibleTests/AndBibleTests/testBuildStrongsMultiDocJSONReturnsInstallFallbackWhenNoStrongsDictionaryIsInstalled -only-testing:AndBibleTests/AndBibleTests/testSpeakServiceMemorizationLoopRepeatsUntilStopped -only-testing:AndBibleTests/AndBibleTests/testSpeakServiceRegularSpeechClearsMemorizationLoop -only-testing:AndBibleTests/AndBibleTests/testSpeakServiceMemorizationLoopIgnoresCancelledReplacedUtterance -only-testing:AndBibleTests/AndBibleTests/testBridgeMemorizationMessagesMutateNativeStore test
- npx gitnexus analyze
- gitnexus detect_changes on staged worktree diff

Impact:
- Reduces BibleReaderController from 9,047 to 8,399 lines on this branch.
- Keeps bridge payload behavior stable while giving future bookmark and StudyPad bridge-action extraction a dedicated payload dependency.
- PR should continue to reference #146 rather than close it until more controller responsibilities are extracted.
